Add envelope interface
The time sliders are set to an exponential detail curve, so that there
is more fine control over envelope settings for lower time settings than
for higher. For example, the difference between a 10ms attack and a 20ms
attack is much more significant and more audible than the difference
between a 4990ms attack and a 5000ms attack.